Use python3 executable instead of system executable
Allows Snakemake rules to use the most appropriate Python 3 executable currently
available in the environment instead of preferring the system executable. This
distinction is most important when users run Snakemake with the `--use-conda`
flag and expect to have access to Python modules installed in the corresponding
conda environment for each rule. When the system executable was preferred, it
would be possible for the rule to run with a Python executable that did not have
access to the same environment as the user's requested conda environment.

Resolves #442